pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aave-point-tracker"
description = "Off-chain Aave liquidity provider point tracker"
authors = [{ name = "András Novoszáth", email = "nocibambi@gmail.com" }]
requires-python = ">=3.13"
readme = "README.md"
dynamic = ["version"]

dependencies = [
    "fastapi[standard]==0.115.5",
    "requests==2.32.3",
    "pandas==2.2.3",
    "python-dotenv==1.0.1",
    "web3==7.5.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel==6.29.5",
    "black==24.10.0",
    "flake8-bugbear==24.10.31",
    "mypy==1.13.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "pytest==8.3.3",
]

[tool.isort]
profile = "black"
